requirements.txt
streamlit
numpy
pandas
streamlit-extras
replicate
python-dotenv
tmdb3
tmdbsimple
